Remove duplicates from mount flags
This patch enhances the mount portion of the block package by making
sure that when mount flags are constructed, duplicates are moved. It
also now makes GetMounts() return all mounts in /proc/mounts, rather
than just those prefixed with /dev. This makes it more useful to callers
who may be looking for mounts other than block devices. This is
preparation to factor the mounts portion of this package out onto its
own.