Add support for BlockVolume type
When a request to publish a BlockVolume comes in, we simply bind mount
the device to the target, with no intermediate "private" mount. A
requirement for this is the target_path has to be a file instead of a
directory.

The Unpublish logic has changed to look at all mounts in the mount
table, and determine if there are any mounts of the device to the target
path, or the private mount area. Looking for mounts of the device to the
target path is different for blockVolume vs MountVolume, but this is
accounted for. Any and all of the three scenarios will be looked for,
and removed accordingly. Private mount can only be removed if the *only*
mount of the device is to the private mount area.